+++ CHANGELOG.md
@@ -27,11 +27,13 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 
 ### Features
 
-- **core**: Added `grab_pointer` to grabs all the pointer input. (#pr @wjian23)
+- **core**: Added `grab_pointer` to grabs all the pointer input. (#669 @wjian23)
+- **widgets**: Added the widget of Slider (#669 @wjian23)
 
 ### Fixed
-- **core**: fix mismatch of providers. (#pr @wjian23)
-- **core**: Added DeclarerWithSubscription to let Widget `Expanded` accept pipe value. (#pr @wjian23)
+
+- **core**: fix mismatch of providers. (#669 @wjian23)
+- **core**: Added DeclarerWithSubscription to let Widget `Expanded` accept pipe value. (#669 @wjian23)
 
 ## [0.4.0-alpha.18] - 2024-12-11